requirements.txt
selenium==4.35.0
python-dotenv==1.1.1
webdriver-manager==4.0.2